+++ src/types/issue.d.ts
@@ -1,0 +1,9 @@
+export type addonData = {
+    name: string,
+    version: string
+};
+
+export type issueData = {
+    addon: addonData;
+    sellerMarketPlaceAlias: string,
+};
