+++ src/types/service.d.ts
@@ -1,0 +1,10 @@
+export type svcResponse<T> = {
+    success: boolean,
+    body?: T,
+    error?: {
+        input: Error | string, options?: {
+            code?: string;
+            exit?: number;
+        }
+    },
+}
